projects/portfolio/06-AI-Chatbot/requirements.txt
Flask==2.3.3
openai==0.28.1
langchain==0.0.310
python-dotenv==1.0.0
flask-cors==4.0.0
